requirements.txt
imap-tools>=1.4.0
python-telegram-bot>=21.0
PyYAML>=6.0
